Get all projects, even if spread over multiple pages
toggl has a default page size of 151 projects in its response. Also there is no clear indication that we miss some results (e.g. a `count` field of available pages, a `next` URL for the next page, ...).
So we now go from page to page on our own and collect all projects.